requirements.txt
python-speech-features
numpy
scipy
sklearn
wave
matplotlib
pyyaml
pandas
torch
torchvision
tqdm